# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Condominios API",
        "description": "API para administración de condominios.\n\nCredenciales locales para probar en Swagger:\n- Email: swagger.admin@example.com\n- Password: Swagger123!\n\nUsa /api/auth/login para obtener el access_token y luego el botón Authorize con el valor Bearer {access_token}.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor actual"
        }
    ],
    "paths": {
...
        "/api/menus": {
            "get": {
                "tags": [
                    "Menús"
                ],
                "summary": "Listar menús",
                "operationId": "menusIndex",
                "responses": {
                    "200": {
                        "description": "Menús encontrados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Menús"
                ],
                "summary": "Crear menú padre o hijo",
                "operationId": "menusStore",
                "responses": {
                    "201": {
                        "description": "Menú creado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/menu": {
            "get": {
                "tags": [
                    "Menús"
                ],
                "summary": "Obtener menú del usuario autenticado",
                "operationId": "authMenu",
                "responses": {
                    "200": {
                        "description": "Menú obtenido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiToken": {
                "type": "http",
                "bearerFormat": "API token",
                "scheme": "bearer"
            }
        }
    },
...
}